+++ tasks_service/add_tasks.go
@@ -1,0 +1,71 @@
+package tasks
+
+import (
+	"database/sql"
+	"encoding/json"
+	"fmt"
+	"net/http"
+	"strconv"
+	"time"
+)
+
+// Обработчик добавления задачи
+func addTaskHandler(w http.ResponseWriter, r *http.Request, db *sql.DB) {
+	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
+
+	var task Task
+	decoder := json.NewDecoder(r.Body)
+	err := decoder.Decode(&task)
+	if err != nil {
+		http.Error(w, `{"error":"Ошибка декодирования JSON"}`, http.StatusBadRequest)
+		return
+	}
+
+	if task.Title == "" {
+		http.Error(w, `{"error":"Не указан заголовок"}`, http.StatusBadRequest)
+		return
+	}
+
+	if task.Date == "" {
+		task.Date = time.Now().Format(dateFormat)
+	} else {
+		_, err := time.Parse(dateFormat, task.Date)
+		if err != nil {
+			http.Error(w, `{"error":"Неверный формат даты"}`, http.StatusBadRequest)
+			return
+		}
+	}
+
+	now := time.Now()
+	tDate, _ := time.Parse(dateFormat, task.Date)
+	tDate, now_fix := compareDates(tDate, now)
+	if tDate.Before(now_fix) {
+		if task.Repeat == "" {
+			task.Date = now.Format(dateFormat)
+		} else {
+			nextDate, err := NextDate(now, task.Date, task.Repeat)
+			if err != nil {
+				http.Error(w, fmt.Sprintf(`{"error":"%s"}`, err.Error()), http.StatusBadRequest)
+				return
+			}
+			task.Date = nextDate
+		}
+	}
+	query := `INSERT INTO scheduler (date, title, comment, repeat) VALUES (?, ?, ?, ?)`
+	res, err := db.Exec(query, task.Date, task.Title, task.Comment, task.Repeat)
+	if err != nil {
+		http.Error(w, fmt.Sprintf(`{"error":"Ошибка добавления: %s"}`, err.Error()), http.StatusInternalServerError)
+		return
+	}
+
+	id, err := res.LastInsertId()
+	if err != nil {
+		http.Error(w, fmt.Sprintf(`{"error":"Ошибка получения id: %s"}`, err.Error()), http.StatusInternalServerError)
+		return
+	}
+
+	task.ID = strconv.Itoa(int(id))
+
+	w.WriteHeader(http.StatusOK)
+	json.NewEncoder(w).Encode(task)
+}
